Eliminate potential deadlock from SIGCHLD signal handler
There were two potential bugs in the way that the SIGCHLD handler was written
previously. In one case, the mutex would be locked when the signal was handled,
and the handler could not get a lock. In other cases, some of the operations
we were doing in the handler were not strictly safe.

As a fix, I have introduced a queue that is populated by the handler and
consumed via a worker thread. These use a semaphore to signal (which is safe).